rsx: Refactor rsx_utils header into smaller pieces
