+++ rpcs3/Emu/Cell/Modules/cellAvconfExt.cpp
@@ -2,7 +2,7 @@
 #include "Emu/system_config.h"
 #include "Emu/Cell/PPUModule.h"
 #include "Emu/IdManager.h"
-#include "Emu/RSX/rsx_utils.h"
+#include "Emu/RSX/Utils/rsx_utils.h"
 #include "Utilities/StrUtil.h"
 
 #include "cellMic.h"
