+++ rpcs3/Emu/RSX/Core/RSXContext.h
@@ -3,7 +3,7 @@
 #include <util/types.hpp>
 #include "Emu/Cell/lv2/sys_rsx.h"
 #include "Emu/RSX/GCM.h"
-#include "Emu/RSX/rsx_utils.h"
+#include "Emu/RSX/Utils/rsx_utils.h"
 #include "RSXIOMap.hpp"
 
 namespace rsx
