+++ rpcs3/Emu/RSX/RSXThread.h
@@ -6,13 +6,13 @@
 #include "RSXFIFO.h"
 #include "RSXOffload.h"
 #include "RSXZCULL.h"
-#include "rsx_utils.h"
 #include "Common/bitfield.hpp"
 #include "Common/profiling_timer.hpp"
 #include "Common/texture_cache_types.h"
 #include "Common/TextureUtils.h"
 #include "Program/RSXVertexProgram.h"
 #include "Program/RSXFragmentProgram.h"
+#include "Utils/rsx_utils.h"
 
 #include "Utilities/Thread.h"
 #include "Utilities/geometry.h"
