+++ rpcs3/Emu/RSX/Utils/color_utils.hpp
@@ -2,7 +2,7 @@
 
 #include <util/types.hpp>
 #include <Utilities/geometry.h>
-#include "gcm_enums.h"
+#include "../gcm_enums.h"
 #include "Utilities/StrFmt.h"
 
 namespace rsx
