+++ rpcs3/Emu/RSX/Utils/rsx_utils.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 #include "rsx_utils.h"
-#include "rsx_methods.h"
+#include "Emu/RSX/rsx_methods.h"
 #include "Emu/Cell/Modules/cellVideoOut.h"
 
 #ifdef _MSC_VER
