+++ rpcs3/Emu/RSX/VK/VKCompute.h
@@ -1,9 +1,10 @@
 #pragma once
-#include "Emu/RSX/VK/VKProgramPipeline.h"
+#include "VKProgramPipeline.h"
 #include "vkutils/descriptors.h"
 #include "vkutils/buffer_object.h"
 
 #include "Emu/IdManager.h"
+#include "Emu/RSX/Utils/algorithm.hpp"
 
 #include "Utilities/StrUtil.h"
 #include "util/asm.hpp"
